requirement.txt
fastapi
uvicorn
numpy
faiss-cpu
requests
python-dotenv
